speedup relation candidates api
